requirements.txt
requests
python-dotenv
Unidecode
setuptools